strudel_sheetmusic/src/omr/omr_engine.py
"""
OMR Engine Module
Core optical music recognition functionality
"""

from typing import Dict, Optional
from PIL import Image
from pathlib import Path


class OMREngine:
    """
    Optical Music Recognition engine
    Converts sheet music images to MusicXML
    """

    def __init__(self, engine: str = "audiveris"):
        """
        Initialize OMR engine

        Args:
            engine: OMR engine to use ('audiveris', 'custom', 'music21')
        """
        self.engine = engine
        self.confidence_scores = {}

    def recognize(
        self,
        image: Image.Image,
        output_path: Optional[str] = None
    ) -> str:
        """
        Recognize music notation from image

        Args:
            image: Input sheet music image
            output_path: Optional path to save MusicXML output

        Returns:
            MusicXML string

        TODO:
        - Implement Audiveris integration
        - Add custom ML model option
        - Implement confidence scoring
        - Add manual correction interface hooks
        """
        raise NotImplementedError("OMR engine not yet implemented")

    def get_confidence_scores(self) -> Dict[str, float]:
        """
        Get confidence scores for last recognition

        Returns:
            Dictionary of element types to confidence scores
        """
        return self.confidence_scores

    def validate_output(self, musicxml: str) -> bool:
        """
        Validate MusicXML output

        Args:
            musicxml: MusicXML string to validate

        Returns:
            True if valid, False otherwise
        """
        # TODO: Implement MusicXML validation
        return True

    def batch_recognize(
        self,
        images: list[Image.Image],
        output_dir: str
    ) -> list[str]:
        """
        Process multiple images in batch

        Args:
            images: List of images to process
            output_dir: Directory to save outputs

        Returns:
            List of MusicXML strings
        """
        results = []
        for i, image in enumerate(images):
            output_path = Path(output_dir) / f"page_{i+1:03d}.musicxml"
            result = self.recognize(image, str(output_path))
            results.append(result)
        return results
