strudel_sheetmusic/requirements.txt
# Core Dependencies
python>=3.9

# PDF and Image Processing
pdf2image>=1.16.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Music Processing
music21>=9.1.0
mido>=1.3.0
python-rtmidi>=1.5.0

# OMR and Machine Learning
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# Web Framework (API)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Async Processing
celery>=5.3.0
redis>=5.0.0

# Data Validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Advanced OMR
# tensorflow>=2.14.0
# torch>=2.1.0
# torchvision>=0.16.0

# Optional: Audiveris integration
# py4j>=0.10.9.7
